+++ src/Libraries/Microsoft.Extensions.AI.OpenAI/OpenAIAssistantsChatClient.cs
@@ -286,7 +286,7 @@ internal sealed class OpenAIAssistantsChatClient : IChatClient
     }
 
     /// <summary>Converts an Extensions function to an OpenAI assistants function tool.</summary>
-    internal static FunctionToolDefinition ToOpenAIAssistantsFunctionToolDefinition(AIFunction aiFunction, ChatOptions? options = null)
+    internal static FunctionToolDefinition ToOpenAIAssistantsFunctionToolDefinition(AIFunctionDeclaration aiFunction, ChatOptions? options = null)
     {
         bool? strict =
             OpenAIClientExtensions.HasStrict(aiFunction.AdditionalProperties) ??
@@ -348,7 +348,7 @@ internal sealed class OpenAIAssistantsChatClient : IChatClient
                 {
                     switch (tool)
                     {
-                        case AIFunction aiFunction:
+                        case AIFunctionDeclaration aiFunction:
                             runOptions.ToolsOverride.Add(ToOpenAIAssistantsFunctionToolDefinition(aiFunction, options));
                             break;